Include racra-wwise
Worked on a set of Python scripts that make it feasible to extract the Wwise soundbanks from the PC port of Rift Apart for PC, as well as a script that tries to replace music specifically. It's niche and rough around the edges, so no worries if this needs to be declined.